style(jobs): move jobs in designated files
